refactor:organization dashboard api
