+++ libs/common/src/interfaces/organization.interface.ts
@@ -1,0 +1,6 @@
+export interface IOrganizationDashboard {
+    usersCount: number,
+    schemasCount: number,
+    credentialsCount: number,
+    presentationsCount:number
+  }
